Desktop: move update controls to Settings > General
