+++ studio/frontend/src/features/settings/components/update-studio-instructions.tsx
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 // Copyright 2026-present the Unsloth AI Inc. team. All rights reserved. See /studio/LICENSE.AGPL-3.0
 
-import { Button } from "@/components/ui/button";
-import { useTauriUpdateController } from "@/hooks/tauri-update-context";
 import { copyToClipboard } from "@/lib/copy-to-clipboard";
 import { useT } from "@/i18n";
 import { Tick02Icon } from "@/lib/tick-icon";
@@ -15,7 +13,6 @@ import { HugeiconsIcon } from "@hugeicons/react";
 import { AnimatePresence, motion, useReducedMotion } from "motion/react";
 import type { ReactElement } from "react";
 import { useEffect, useRef, useState } from "react";
-import { SettingsRow } from "./settings-row";
 
 const STUDIO_INSTALL_UNIX_CMD =
   "curl -fsSL https://unsloth.ai/install.sh | sh";
@@ -175,87 +172,6 @@ function UpdateDocsLinks(): ReactElement {
   );
 }
 
-function formatDesktopVersion(version: string): string {
-  return version.startsWith("v") ? version : `v${version}`;
-}
-
-function DesktopUpdateControl(): ReactElement | null {
-  const t = useT();
-  const update = useTauriUpdateController();
-  if (!update) return null;
-
-  const checking = update.status === "checking";
-  // A running install owns the update screen; no second "Update now".
-  const inFlight =
-    update.status === "updating-backend" ||
-    update.status === "downloading" ||
-    update.status === "installing";
-  const busy = checking || inFlight;
-  const available = update.info !== null && !checking;
-  const checkFailed = update.checkError !== null && !available;
-
-  let label: string;
-  let description: string;
-  if (checking) {
-    label = t("settings.about.update.desktopChecking");
-    description = t("settings.about.update.desktopCheckingDescription");
-  } else if (available && update.info) {
-    label = t("settings.about.update.desktopAvailable", {
-      version: formatDesktopVersion(update.info.version),
-    });
-    description = update.isExternalServer
-      ? t("settings.about.update.desktopExternalServer")
-      : update.updatePolicyMode === "manual_linux_package"
-        ? t("settings.about.update.desktopManualInstall")
-        : t("settings.about.update.desktopAvailableDescription");
-  } else if (checkFailed) {
-    label = t("settings.about.update.desktopCheckFailed");
-    // Keep the reason: it separates offline from a bad release manifest.
-    description = `${t("settings.about.update.desktopCheckFailedDescription")} ${update.checkError}`;
-  } else if (update.hasChecked) {
-    label = t("settings.about.update.desktopCurrent");
-    description = t("settings.about.update.desktopCurrentDescription");
-  } else {
-    label = t("settings.about.update.desktopReady");
-    description = t("settings.about.update.desktopReadyDescription");
-  }
-
-  const action = available
-    ? update.updatePolicyMode === "manual_linux_package"
-      ? t("settings.about.update.openReleasePage")
-      : t("settings.about.update.updateNow")
-    : checkFailed
-      ? t("settings.about.update.retryCheck")
-      : update.hasChecked
-        ? t("settings.about.update.checkAgain")
-        : t("settings.about.update.checkForUpdates");
-
-  return (
-    <SettingsRow
-      label={label}
-      description={<span aria-live="polite">{description}</span>}
-    >
-      {available && update.isExternalServer ? null : (
-        <Button
-          size="sm"
-          variant={available ? "default" : "outline"}
-          disabled={busy}
-          aria-busy={busy}
-          onClick={() => {
-            if (available) {
-              void update.installUpdate();
-            } else {
-              void update.checkForUpdate();
-            }
-          }}
-        >
-          {checking ? t("settings.about.update.checking") : action}
-        </Button>
-      )}
-    </SettingsRow>
-  );
-}
-
 function ShellToggleButton({
   active,
   label,
@@ -328,10 +244,8 @@ export function UpdateStudioInstructions({
             {t("settings.about.update.title")}
           </p>
         ) : null}
-        <p className="text-xs text-muted-foreground leading-relaxed">
-          {t("settings.about.update.desktopManaged")}
-        </p>
-        <DesktopUpdateControl />
+        {/* The check/update control and its explanation live in General, next
+            to the version rows, so there is exactly one place to update from. */}
         <UpdateDocsLinks />
       </div>
     );
