+++ tests/studio/test_desktop_reliability_frontend_contract.py
@@ -25,6 +25,7 @@ TAURI_MAIN = REPO / "studio/src-tauri/src/main.rs"
 TAURI_UPDATE_CONTEXT = FRONTEND / "hooks/tauri-update-context.ts"
 TAURI_UPDATE_HOOK = FRONTEND / "hooks/use-tauri-update.ts"
 UPDATE_INSTRUCTIONS = FRONTEND / "features/settings/components/update-studio-instructions.tsx"
+DESKTOP_UPDATE_CONTROL = FRONTEND / "features/settings/components/desktop-update-control.tsx"
 DESKTOP_UPDATE_POLICY = REPO / "studio/src-tauri/src/desktop_update_policy.rs"
 
 
@@ -43,7 +44,7 @@ def test_desktop_update_offer_remains_actionable_from_settings():
     provider = APP_PROVIDER.read_text(encoding = "utf-8")
     context = TAURI_UPDATE_CONTEXT.read_text(encoding = "utf-8")
     hook = TAURI_UPDATE_HOOK.read_text(encoding = "utf-8")
-    settings = UPDATE_INSTRUCTIONS.read_text(encoding = "utf-8")
+    settings = DESKTOP_UPDATE_CONTROL.read_text(encoding = "utf-8")
 
     assert "<TauriUpdateContext.Provider value={update}>" in provider
     context_start = provider.index("<TauriUpdateContext.Provider value={update}>")
@@ -92,7 +93,7 @@ def test_desktop_update_keeps_the_in_app_path_on_a_guessed_policy():
 
 
 def test_settings_update_button_is_inert_while_an_install_runs():
-    settings = UPDATE_INSTRUCTIONS.read_text(encoding = "utf-8")
+    settings = DESKTOP_UPDATE_CONTROL.read_text(encoding = "utf-8")
 
     assert 'update.status === "updating-backend"' in settings
     assert 'update.status === "downloading"' in settings
@@ -103,7 +104,7 @@ def test_settings_update_button_is_inert_while_an_install_runs():
 
 def test_desktop_update_check_failures_are_retryable():
     hook = TAURI_UPDATE_HOOK.read_text(encoding = "utf-8")
-    settings = UPDATE_INSTRUCTIONS.read_text(encoding = "utf-8")
+    settings = DESKTOP_UPDATE_CONTROL.read_text(encoding = "utf-8")
     policy = DESKTOP_UPDATE_POLICY.read_text(encoding = "utf-8")
 
     assert "setCheckError(String(e));" in hook
